docker working solution
