+++ src/services/simulation_service/core/infrastructure/server/src/_simulation_server_grpc.py
@@ -10,7 +10,7 @@ import services.simulation_service.core.infrastructure.generated.simulation_mess
 import services.simulation_service.core.infrastructure.generated.simulation_messaging_pb2_grpc as sm_grpc
 from logger import get_logger
 
-logger = get_logger("threading-server")
+logger = get_logger("threading-server", filename=__name__)
 
 SERVER_OPTIONS = [
     ("grpc.max_send_message_length", 100 * 1024 * 1024),  # 100MB
